Change shape of LINQ support.
Renamed "Query" namespaces to "Linq".

Language version bump to 7.1.

Moved LINQ support of Opt to new LineToOpt class.

Fixed doubling of Convert and Try dirs.
