+++ Hgk.Zero.Options/Linq/LinqToOpt_Zip.cs
@@ -1,0 +1,35 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+
+namespace Hgk.Zero.Options.Linq
+{
+    public static partial class LinqToOpt
+    {
+        /// <inheritdoc cref="Enumerable.Zip{TFirst, TSecond, TResult}(IEnumerable{TFirst}, IEnumerable{TSecond}, Func{TFirst, TSecond, TResult})"/>
+        /// <remarks>
+        /// <para>
+        /// This method is implemented using deferred execution; the query represented by this method
+        /// is not performed until the contents of the returned option are resolved, such as by enumeration.
+        /// </para>
+        /// </remarks>
+        public static IOpt<TResult> Zip<TFirst, TSecond, TResult>(this IOpt<TFirst> first, IEnumerable<TSecond> second, Func<TFirst, TSecond, TResult> resultSelector)
+        {
+            if (first == null) throw new ArgumentNullException(nameof(first));
+            if (second == null) throw new ArgumentNullException(nameof(second));
+            if (resultSelector == null) throw new ArgumentNullException(nameof(resultSelector));
+
+            return first.MetaSelect(opt =>
+            {
+                if (opt.HasValue)
+                {
+                    foreach (var secondValue in second)
+                    {
+                        return Opt.Full(resultSelector(opt.ValueOrDefault, secondValue));
+                    }
+                }
+                return Opt.Empty<TResult>();
+            });
+        }
+    }
+}
