+++ Hgk.Zero.Options/Opt_basic.cs
@@ -1,49 +1,13 @@
 ﻿using System;
-using System.Collections;
 using System.Collections.Generic;
-using System.Linq;
 
 namespace Hgk.Zero.Options
 {
     /// <summary>
     /// Facilities to create and manipulate options.
     /// </summary>
-    /// <remarks>
-    /// <para>
-    /// This class contains a large number of LINQ-compatible methods. While all options are also
-    /// enumerables, and can therefore be used with methods provided by <see cref="Enumerable"/>,
-    /// most of the operators are specialized here for options in order to optimize the algorithms
-    /// for use with zero- or one-element collections (such as <see
-    /// cref="Opt.Distinct{TSource}(IOpt{TSource})"/>), and also to return options from operations
-    /// guaranteed to return at most one element when performed on an option (such as <see
-    /// cref="Opt.Zip{TFirst, TSecond, TResult}(IOpt{TFirst}, IEnumerable{TSecond}, Func{TFirst,
-    /// TSecond, TResult})"/>). In general, the result of applying any LINQ operation in this class
-    /// to an option is defined to be sequence-equal to the result of applying the same operation
-    /// from <see cref="Enumerable"/> to the same option.
-    /// </para>
-    /// <para>
-    /// Similarly to the methods in <see cref="Enumerable"/>, most option-returning methods in this
-    /// class are implemented using deferred execution; that is, the query represented by the method
-    /// is not actually executed until the contents of the result are resolved directly, for example,
-    /// by enumeration, using foreach, or calling a method such as <see
-    /// cref="ToFixed{TSource}(IOpt{TSource})"/> or <see cref="Match{TSource,
-    /// TResult}(IOpt{TSource}, Func{TResult}, Func{TSource, TResult})"/>.
-    /// </para>
-    /// <para>
-    /// Some methods instead use or return <see cref="Opt{T}"/>, which is a type for fixed, immutable
-    /// options. The contents of these are immediate and not deferred.
-    /// </para>
-    /// </remarks>
     public static partial class Opt
     {
-        /// <summary>
-        /// Returns this option typed as <see cref="IOpt{T}"/>.
-        /// </summary>
-        /// <typeparam name="TSource">The element type of <paramref name="source"/>.</typeparam>
-        /// <param name="source">An option to type as <see cref="IOpt{T}"/>.</param>
-        /// <returns><paramref name="source"/>, typed as <see cref="IOpt{T}"/>.</returns>
-        public static IOpt<TSource> AsIOpt<TSource>(this IOpt<TSource> source) => source;
-
         /// <summary>
         /// Creates a fixed option with the same contents as the specified <see cref="Nullable{T}"/>.
         /// </summary>
@@ -93,35 +57,6 @@ namespace Hgk.Zero.Options
         public static Opt<T> Create<T>(bool hasValue, T value) => new Opt<T>(hasValue, value);
 
         /// <summary>
-        /// Returns the element at a specified index in an option or a default value if the index is
-        /// out of range.
-        /// </summary>
-        /// <typeparam name="TSource">The type of the elements of <paramref name="source"/>.</typeparam>
-        /// <param name="source">An option to return an element from.</param>
-        /// <param name="index">The index of the element to retrieve.</param>
-        /// <param name="defaultValue">
-        /// A default value to return if <paramref name="index"/> is out of bounds.
-        /// </param>
-        /// <returns>
-        /// <paramref name="defaultValue"/> if the index is outside the bounds of <paramref
-        /// name="source"/>; otherwise, the element at the specified position in <paramref name="source"/>.
-        /// </returns>
-        /// <exception cref="ArgumentNullException"><paramref name="source"/> is <see langword="null"/>.</exception>
-        public static TSource ElementAtOrDefault<TSource>(this IOpt<TSource> source, int index, TSource defaultValue)
-        {
-            if (source == null) throw new ArgumentNullException(nameof(source));
-            if (index == 0)
-            {
-                var opt = source.ToFixed();
-                if (opt.HasValue)
-                {
-                    return opt.ValueOrDefault;
-                }
-            }
-            return defaultValue;
-        }
-
-        /// <summary>
         /// Creates a fixed, empty option.
         /// </summary>
         /// <typeparam name="T">The element type of the new option.</typeparam>
@@ -133,284 +68,12 @@ namespace Hgk.Zero.Options
         public static Opt<T> Empty<T>() => new Opt<T>();
 
         /// <summary>
-        /// Returns the first element of an option, or a default value if the option contains no elements.
-        /// </summary>
-        /// <typeparam name="TSource">The type of the elements of <paramref name="source"/>.</typeparam>
-        /// <param name="source">An option to return an element from.</param>
-        /// <param name="defaultValue">A default value to return if <paramref name="source"/> contains no elements.</param>
-        /// <returns><paramref name="defaultValue"/> if <paramref name="source"/> is empty; otherwise, the first element in <paramref name="source"/>.</returns>
-        /// <exception cref="ArgumentNullException"><paramref name="source"/> is <see langword="null"/>.</exception>
-        public static TSource FirstOrDefault<TSource>(this IOpt<TSource> source, TSource defaultValue) => source.SingleOrDefault(defaultValue);
-
-        /// <summary>
-        /// Returns the first element of the option that satisfies a condition or a default value if no such element is found.
-        /// </summary>
-        /// <typeparam name="TSource">The type of the elements of <paramref name="source"/>.</typeparam>
-        /// <param name="source">An option to return an element from.</param>
-        /// <param name="predicate">A function to test each element for a condition.</param>
-        /// <param name="defaultValue">A default value to return if <paramref name="source"/> contains no elements that satisfy <paramref name="predicate"/>.</param>
-        /// <returns><paramref name="defaultValue"/> if <paramref name="source"/> is empty or if no element passes the test specified by <paramref name="predicate"/>; otherwise, the first element in <paramref name="source"/> that passes the test specified by <paramref name="predicate"/>.</returns>
-        /// <exception cref="ArgumentNullException"><paramref name="source"/> is <see langword="null"/>.</exception>
-        public static TSource FirstOrDefault<TSource>(this IOpt<TSource> source, Func<TSource, bool> predicate, TSource defaultValue) => source.SingleOrDefault(predicate, defaultValue);
-
-        /// <summary>
-        /// Collapses an option containing another option into a single option.
-        /// </summary>
-        /// <remarks>
-        /// <para>
-        /// This method is implemented using deferred execution; the query represented by this method
-        /// is not performed until the contents of the returned option are resolved, such as by enumeration.
-        /// </para>
-        /// </remarks>
-        /// <typeparam name="TSource">The element type of the inner option.</typeparam>
-        /// <param name="source">An outer option containing an inner option option.</param>
-        /// <returns>
-        /// An option equivalent to the inner option, if it exists; otherwise, an empty option.
-        /// </returns>
-        public static IOpt<TSource> Flatten<TSource>(this IOpt<IOpt<TSource>> source)
-        {
-            if (source == null) throw new ArgumentNullException(nameof(source));
-            return source.MetaSelect(FlattenFixedOpt);
-        }
-
-        /// <summary>
-        /// Collapses an option containing another option into a single option.
-        /// </summary>
-        /// <remarks>
-        /// <para>
-        /// This method is implemented using deferred execution; the query represented by this method
-        /// is not performed until the contents of the returned option are resolved, such as by enumeration.
-        /// </para>
-        /// </remarks>
-        /// <typeparam name="TSource">The element type of the inner option.</typeparam>
-        /// <param name="source">An outer option containing an inner option option.</param>
-        /// <returns>
-        /// An option equivalent to the inner option, if it exists; otherwise, an empty option.
-        /// </returns>
-        public static IOpt<TSource> Flatten<TSource>(this IOpt<Opt<TSource>> source)
-        {
-            if (source == null) throw new ArgumentNullException(nameof(source));
-            return source.MetaSelect(FlattenFixedOpt);
-        }
-
-        /// <summary>
-        /// Creates a fixed option that contains the specified value (even if the value is <see langword="null"/>).
-        /// </summary>
-        /// <typeparam name="T">The type of <paramref name="value"/>.</typeparam>
-        /// <param name="value">The value to be contained in the new option.</param>
-        /// <returns>An option that contains <paramref name="value"/>.</returns>
-        /// <seealso cref="Empty{T}"/>
-        /// <seealso cref="Create{T}(T)"/>
-        /// <seealso cref="Create{T}(T?)"/>
-        /// <seealso cref="Create{T}(bool, T)"/>
-        public static Opt<T> Full<T>(T value) => new Opt<T>(true, value);
-
-        /// <summary>
-        /// Returns the last element of an option, or a default value if the option contains no elements.
-        /// </summary>
-        /// <typeparam name="TSource"></typeparam>
-        /// <param name="source">An option to return an element from.</param>
-        /// <param name="defaultValue">A default value to return if <paramref name="source"/> contains no elements.</param>
-        /// <returns><paramref name="defaultValue"/> if <paramref name="source"/> is empty; otherwise, the last element in <paramref name="source"/>.</returns>
-        /// <exception cref="ArgumentNullException"><paramref name="source"/> is <see langword="null"/>.</exception>
-        public static TSource LastOrDefault<TSource>(this IOpt<TSource> source, TSource defaultValue) => source.SingleOrDefault(defaultValue);
-
-        /// <summary>
-        /// Returns the last element of an option that satisfies a condition or a default value if no such element is found.
-        /// </summary>
-        /// <typeparam name="TSource"></typeparam>
-        /// <param name="source">An option to return an element from.</param>
-        /// <param name="predicate">A function to test each element for a condition.</param>
-        /// <param name="defaultValue">A default value to return if <paramref name="source"/> contains no elements that satisfy <paramref name="predicate"/>.</param>
-        /// <returns><paramref name="defaultValue"/> if the option is empty or if no elements pass the test in the predicate function; otherwise, the last element that passes the test in the predicate function.</returns>
-        /// <exception cref="ArgumentNullException"><paramref name="source"/> is <see langword="null"/>.</exception>
-        public static TSource LastOrDefault<TSource>(this IOpt<TSource> source, Func<TSource, bool> predicate, TSource defaultValue) => source.SingleOrDefault(predicate, defaultValue);
-
-        /// <summary>
-        /// Converts this option to another value based on its contents.
-        /// </summary>
-        /// <remarks>
-        /// <para>
-        /// This operation is performed immediately rather than by deferred execution. Compare <see
-        /// cref="MatchOpt"/>, which uses deferred execution.
-        /// </para>
-        /// </remarks>
-        /// <typeparam name="TSource">The element type of <paramref name="source"/>.</typeparam>
-        /// <typeparam name="TResult">The result type for this conversion.</typeparam>
-        /// <param name="source">A source option.</param>
-        /// <param name="ifEmpty">A function to produce the result if this option is empty.</param>
-        /// <param name="ifFull">
-        /// A function to produce the result if this option is full, which accepts the contained
-        /// element as a parameter.
-        /// </param>
-        /// <returns>
-        /// The result of calling <paramref name="ifEmpty"/>, if this option is empty; otherwise, the
-        /// result of calling <paramref name="ifFull"/>.
-        /// </returns>
-        /// <exception cref="ArgumentNullException"><paramref name="source"/> is <see langword="null"/>.</exception>
-        /// <exception cref="InvalidOperationException">
-        /// <paramref name="ifEmpty"/> is <see langword="null"/> and this option is empty, or
-        /// <paramref name="ifFull"/> is <see langword="null"/> and this option is full.
-        /// </exception>
-        public static TResult Match<TSource, TResult>(this IOpt<TSource> source, Func<TResult> ifEmpty = null, Func<TSource, TResult> ifFull = null)
-        {
-            var opt = source.ToFixed();
-            if (opt.HasValue)
-            {
-                if (ifFull == null)
-                {
-                    throw Error.MatchCaseFailed();
-                }
-                return ifFull(opt.ValueOrDefault);
-            }
-            else
-            {
-                if (ifEmpty == null)
-                {
-                    throw Error.MatchCaseFailed();
-                }
-                return ifEmpty();
-            }
-        }
-
-        /// <summary>
-        /// Converts this option to another option based on its contents.
-        /// </summary>
-        /// <remarks>
-        /// <para>
-        /// This method is implemented using deferred execution; the query represented by this method
-        /// is not performed until the contents of the returned option are resolved, such as by enumeration.
-        /// </para>
-        /// <para>Compare <see cref="Match"/>, which uses immediate execution.</para>
-        /// </remarks>
-        /// <typeparam name="TSource">The element type of <paramref name="source"/>.</typeparam>
-        /// <typeparam name="TResult">The element type for the result of this conversion.</typeparam>
-        /// <param name="source">A source option.</param>
-        /// <param name="ifEmpty">A function to produce the result if this option is empty.</param>
-        /// <param name="ifFull">
-        /// A function to produce the result if this option is full, which accepts the contained
-        /// element as a parameter.
-        /// </param>
-        /// <returns>
-        /// An option equivalent to the result of calling <paramref name="ifEmpty"/>, if this option
-        /// is empty; otherwise, an option equivalent to the result of calling <paramref name="ifFull"/>.
-        /// </returns>
-        /// <exception cref="ArgumentNullException"><paramref name="source"/> is <see langword="null"/>.</exception>
-        /// <exception cref="InvalidOperationException">
-        /// <paramref name="ifEmpty"/> is <see langword="null"/> and this option is empty, or
-        /// <paramref name="ifFull"/> is <see langword="null"/> and this option is full.
-        /// </exception>
-        public static IOpt<TResult> MatchOpt<TSource, TResult>(this IOpt<TSource> source, Func<Opt<TResult>> ifEmpty = null, Func<TSource, Opt<TResult>> ifFull = null)
-        {
-            if (source == null) throw new ArgumentNullException(nameof(source));
-            return Defer(() => source.Match(ifEmpty, ifFull));
-        }
-
-        /// <summary>
-        /// Substitutes the result of a callback if this option is empty.
-        /// </summary>
-        /// <typeparam name="TSource">The element type of <paramref name="source"/>.</typeparam>
-        /// <param name="source">A source option.</param>
-        /// <param name="replacementFactory">
-        /// A function to produce the result if <paramref name="source"/> is empty.
-        /// </param>
-        /// <returns>
-        /// An option equivalent to <paramref name="source"/>, if it is full;
-        /// otherwise, an option equivalent to the result of calling <paramref name="replacementFactory"/>.
-        /// </returns>
-        /// <exception cref="ArgumentNullException">
-        /// <paramref name="source"/> or <paramref name="replacementFactory"/> is <see langword="null"/>.
-        /// </exception>
-        public static IOpt<TSource> ReplaceIfEmpty<TSource>(this IOpt<TSource> source, Func<IOpt<TSource>> replacementFactory)
-        {
-            if (source == null) throw new ArgumentNullException(nameof(source));
-            if (replacementFactory == null) throw new ArgumentNullException(nameof(replacementFactory));
-            return Defer(() => source.ToFixed().ReplaceIfEmptyImmediate(replacementFactory));
-        }
-
-        /// <summary>
-        /// Substitutes the specified replacement option if this option is empty.
-        /// </summary>
-        /// <typeparam name="TSource">The element type of <paramref name="source"/>.</typeparam>
-        /// <param name="source">A source option.</param>
-        /// <param name="replacement">
-        /// An option to produce the result if <paramref name="source"/> is empty.
-        /// </param>
-        /// <returns>
-        /// An option equivalent to <paramref name="source"/>, if it is full; otherwise, an option equivalent to <paramref name="replacement"/>.
-        /// </returns>
-        /// <exception cref="ArgumentNullException">
-        /// <paramref name="source"/> or <paramref name="replacement"/> is <see langword="null"/>.
-        /// </exception>
-        public static IOpt<TSource> ReplaceIfEmpty<TSource>(this IOpt<TSource> source, IOpt<TSource> replacement)
-        {
-            if (source == null) throw new ArgumentNullException(nameof(source));
-            if (replacement == null) throw new ArgumentNullException(nameof(replacement));
-            return Defer(() => source.ToFixed().ReplaceIfEmptyImmediate(replacement));
-        }
-
-        /// <summary>
-        /// Returns the only element of an option, or a default value if the option is empty.
-        /// </summary>
-        /// <typeparam name="TSource"></typeparam>
-        /// <param name="source">An option to return an element from.</param>
-        /// <param name="defaultValue">A default value to return if <paramref name="source"/> contains no elements.</param>
-        /// <returns>The single element of the input option, or <paramref name="defaultValue"/> if the option contains no elements.</returns>
-        /// <exception cref="ArgumentNullException"><paramref name="source"/> is <see langword="null"/>.</exception>
-        public static TSource SingleOrDefault<TSource>(this IOpt<TSource> source, TSource defaultValue)
-        {
-            var opt = source.ToFixed();
-            return opt.HasValue ? opt.ValueOrDefault : defaultValue;
-        }
-
-        /// <summary>
-        /// Returns the only element of an option that satisfies a specified condition or a default value if no such element exists.
-        /// </summary>
-        /// <typeparam name="TSource"></typeparam>
-        /// <param name="source">An option to return an element from.</param>
-        /// <param name="predicate">A function to test each element for a condition.</param>
-        /// <param name="defaultValue">A default value to return if <paramref name="source"/> contains no elements that satisfy <paramref name="predicate"/>.</param>
-        /// <returns>The single element of the input option that satisfies the condition, or <paramref name="defaultValue"/> if no such element is found.</returns>
-        /// <exception cref="ArgumentNullException"><paramref name="source"/> is <see langword="null"/>.</exception>
-        public static TSource SingleOrDefault<TSource>(this IOpt<TSource> source, Func<TSource, bool> predicate, TSource defaultValue)
-        {
-            if (predicate == null) throw new ArgumentNullException(nameof(predicate));
-            var opt = source.ToFixed();
-            return (opt.HasValue && predicate(opt.ValueOrDefault)) ? opt.ValueOrDefault : defaultValue;
-        }
-
-        /// <summary>
-        /// Returns a fixed option reflecting the current state of an option.
-        /// </summary>
-        /// <param name="source">A source option.</param>
-        /// <returns>A fixed option reflecting the current state of <paramref name="source"/>.</returns>
-        public static Opt<object> ToFixed(this IOpt source)
-        {
-            if (source == null)
-            {
-                throw new ArgumentNullException(nameof(source));
-            }
-            else if (source is Opt<object> alreadyFixedSource)
-            {
-                return alreadyFixedSource;
-            }
-            else if (source is IOptFixable fixableSource)
-            {
-                return fixableSource.ToFixed();
-            }
-            else
-            {
-                return ToFixed(source.GetEnumerator());
-            }
-        }
-
-        /// <summary>
         /// Returns a fixed option reflecting the current state of an option.
         /// </summary>
         /// <typeparam name="TSource">The element type of <paramref name="source"/>.</typeparam>
         /// <param name="source">A source option.</param>
         /// <returns>A fixed option reflecting the current state of <paramref name="source"/>.</returns>
-        public static Opt<TSource> ToFixed<TSource>(this IOpt<TSource> source)
+        public static Opt<TSource> Fix<TSource>(IOpt<TSource> source)
         {
             if (source == null)
             {
@@ -428,82 +91,49 @@ namespace Hgk.Zero.Options
             {
                 using (var se = source.GetEnumerator())
                 {
-                    return ToFixed(se);
+                    return Fix(se);
                 }
             }
         }
 
         /// <summary>
-        /// Gets the value contained by an option, if it exists.
+        /// Returns a fixed option reflecting the current state of an option.
         /// </summary>
-        /// <typeparam name="TSource">The element type of source.</typeparam>
-        /// <param name="source">An option whose value to get.</param>
-        /// <param name="value">
-        /// When this method returns, is set to the value contained by <paramref name="source"/>, if
-        /// any, or the default value of <typeparamref name="TSource"/>, otherwise.
-        /// </param>
-        /// <returns><see langword="true"/>, if source contains a value; otherwise, <see langword="false"/>.</returns>
-        /// <exception cref="ArgumentNullException"><paramref name="source"/> is <see langword="null"/>.</exception>
-        public static bool TryGetValue<TSource>(this IOpt<TSource> source, out TSource value) =>
-            source.ToFixed().TryGetValueRaw(out value);
-
-        /// <summary>
-        /// Filters the contents of an option to include only non- <see langword="null"/> values.
-        /// </summary>
-        /// <remarks>
-        /// <para>
-        /// This method is implemented using deferred execution; the query represented by this method
-        /// is not performed until the contents of the returned option are resolved, such as by enumeration.
-        /// </para>
-        /// </remarks>
-        /// <typeparam name="TSource">The type of the elements of <paramref name="source"/>.</typeparam>
-        /// <param name="source">An option to filter.</param>
-        /// <returns>
-        /// An option equivalent to <paramref name="source"/>, if it is full and contains a non- <see
-        /// langword="null"/> value; otherwise, an empty option.
-        /// </returns>
-        /// <exception cref="ArgumentNullException"><paramref name="source"/> is <see langword="null"/>.</exception>
-        public static IOpt<TSource> WhereNotNull<TSource>(this IOpt<TSource> source)
+        /// <param name="source">A source option.</param>
+        /// <returns>A fixed option reflecting the current state of <paramref name="source"/>.</returns>
+        public static Opt<object> FixUntyped(IOpt source)
         {
-            if (source == null) throw new ArgumentNullException(nameof(source));
-            return source.MetaSelect(opt => opt.WhereNotNullRaw());
-        }
-
-        /// <summary>
-        /// Filters the contents of an option to include only non- <see langword="null"/> values,
-        /// converting the results to their non-nullable type.
-        /// </summary>
-        /// <remarks>
-        /// <para>
-        /// This method is implemented using deferred execution; the query represented by this method
-        /// is not performed until the contents of the returned option are resolved, such as by enumeration.
-        /// </para>
-        /// </remarks>
-        /// <typeparam name="TSource">
-        /// The non-nullable basis type of the type of the elements of <paramref name="source"/>.
-        /// </typeparam>
-        /// <param name="source">An option to filter.</param>
-        /// <returns>
-        /// An option containing the same value as source, if it is full and contains a non- <see
-        /// langword="null"/> value; otherwise, an empty option.
-        /// </returns>
-        /// <exception cref="ArgumentNullException"><paramref name="source"/> is <see langword="null"/>.</exception>
-        public static IOpt<TSource> WhereNotNullSelectValue<TSource>(this IOpt<TSource?> source) where TSource : struct
-        {
-            if (source == null) throw new ArgumentNullException(nameof(source));
-            return source.MetaSelect(opt => opt.WhereNotNullSelectValueRaw());
-        }
-
-        internal static bool Contains<TSource>(Opt<TSource> source, TSource value, IEqualityComparer<TSource> comparer)
-        {
-            if (source.HasValue)
+            if (source == null)
             {
-                return (comparer ?? EqualityComparer<TSource>.Default).Equals(source.ValueOrDefault, value);
+                throw new ArgumentNullException(nameof(source));
             }
-            return false;
+            else if (source is Opt<object> alreadyFixedSource)
+            {
+                return alreadyFixedSource;
+            }
+            else if (source is IOptFixable fixableSource)
+            {
+                return fixableSource.ToFixed();
+            }
+            else
+            {
+                return FixUntyped(source.GetEnumerator());
+            }
         }
 
-        internal static DeferredOpt<TSource> Defer<TSource>(this Func<Opt<TSource>> toFixedFunction)
+        /// <summary>
+        /// Creates a fixed option that contains the specified value (even if the value is <see langword="null"/>).
+        /// </summary>
+        /// <typeparam name="T">The type of <paramref name="value"/>.</typeparam>
+        /// <param name="value">The value to be contained in the new option.</param>
+        /// <returns>An option that contains <paramref name="value"/>.</returns>
+        /// <seealso cref="Empty{T}"/>
+        /// <seealso cref="Create{T}(T)"/>
+        /// <seealso cref="Create{T}(T?)"/>
+        /// <seealso cref="Create{T}(bool, T)"/>
+        public static Opt<T> Full<T>(T value) => new Opt<T>(true, value);
+
+        internal static DeferredOpt<TSource> DeferRaw<TSource>(this Func<Opt<TSource>> toFixedFunction)
         {
             return new DeferredOpt<TSource>(toFixedFunction);
         }
@@ -544,8 +174,8 @@ namespace Hgk.Zero.Options
             }
             else
             {
-                var fixedA = a.ToFixed();
-                var fixedB = b.ToFixed();
+                var fixedA = FixUntyped(a);
+                var fixedB = FixUntyped(b);
                 if (fixedA.HasValue)
                 {
                     return fixedB.HasValue && Equals(fixedA.ValueOrDefault, fixedB.ValueOrDefault);
@@ -557,106 +187,12 @@ namespace Hgk.Zero.Options
             }
         }
 
-        internal static Opt<TSource> WhereRaw<TSource>(this Opt<TSource> source, Func<TSource, bool> predicate)
-        {
-            return (source.HasValue && predicate(source.ValueOrDefault)) ? source : Empty<TSource>();
-        }
-
-        private static Opt<TResult> Cast<TSource, TResult>(Opt<TSource> source)
-        {
-            return source.HasValue ? Full((TResult)(object)source.ValueOrDefault) : Empty<TResult>();
-        }
-
-        private static Opt<TSource> FlattenFixedOpt<TSource>(Opt<IOpt<TSource>> fixedOpt1)
-        {
-            if (fixedOpt1.HasValue)
-            {
-                var fixedOpt0 = fixedOpt1.ValueOrDefault.ToFixed();
-                return fixedOpt0;
-            }
-            return Empty<TSource>();
-        }
-
-        private static Opt<TSource> FlattenFixedOpt<TSource>(Opt<Opt<TSource>> fixedOpt1)
-        {
-            if (fixedOpt1.HasValue)
-            {
-                var fixedOpt0 = fixedOpt1.ValueOrDefault;
-                return fixedOpt0;
-            }
-            return Empty<TSource>();
-        }
-
-        private static IOpt<TSource> KeepIf<TSource>(this IOpt<TSource> source, bool condition) => condition ? source : Empty<TSource>();
-
-        private static MetaSelectOpt<TSource, TResult> MetaSelect<TSource, TResult>(this IOpt<TSource> source, Func<Opt<TSource>, Opt<TResult>> metaselector)
+        internal static MetaSelectOpt<TSource, TResult> MetaSelect<TSource, TResult>(this IOpt<TSource> source, Func<Opt<TSource>, Opt<TResult>> metaselector)
         {
             return new MetaSelectOpt<TSource, TResult>(source, metaselector);
         }
 
-        private static Opt<TResult> OfType<TResult>(bool hasValue, object valueOrDefault)
-        {
-            return hasValue && (valueOrDefault is TResult value) ? Full(value) : Empty<TResult>();
-        }
-
-        private static Opt<TResult> OfType<TSource, TResult>(Opt<TSource> source)
-        {
-            return OfType<TResult>(source.HasValue, source.ValueOrDefault);
-        }
-
-        private static Opt<TSource> ReplaceIfEmptyImmediate<TSource>(this Opt<TSource> source, Func<IOpt<TSource>> replacementFactory)
-        {
-            return source.HasValue ? source : replacementFactory().ToFixed();
-        }
-
-        private static Opt<TSource> ReplaceIfEmptyImmediate<TSource>(this Opt<TSource> source, IOpt<TSource> replacement)
-        {
-            return source.HasValue ? source : replacement.ToFixed();
-        }
-
-        private static Opt<TResult> SelectManyRaw<TSource, TResult>(this Opt<TSource> source, Func<TSource, IOpt<TResult>> selector)
-        {
-            return FlattenFixedOpt(source.SelectRaw(selector));
-        }
-
-        private static Opt<TResult> SelectManyRaw<TSource, TResult>(this Opt<TSource> source, Func<TSource, int, IOpt<TResult>> selector)
-        {
-            return FlattenFixedOpt(source.SelectRaw(selector));
-        }
-
-        private static Opt<TResult> SelectManyRaw<TSource, TCollection, TResult>(this Opt<TSource> source, Func<TSource, IOpt<TCollection>> collectionSelector, Func<TSource, TCollection, TResult> resultSelector)
-        {
-            return source.SelectManyRaw(element => SubSelect(element, collectionSelector(element), resultSelector));
-        }
-
-        private static Opt<TResult> SelectManyRaw<TSource, TCollection, TResult>(this Opt<TSource> source, Func<TSource, int, IOpt<TCollection>> collectionSelector, Func<TSource, TCollection, TResult> resultSelector)
-        {
-            return source.SelectManyRaw(element => SubSelect(element, collectionSelector(element, 0), resultSelector));
-        }
-
-        private static Opt<TResult> SelectRaw<TSource, TResult>(this Opt<TSource> source, Func<TSource, TResult> selector)
-        {
-            return source.HasValue ? Full(selector(source.ValueOrDefault)) : Empty<TResult>();
-        }
-
-        private static Opt<TResult> SelectRaw<TSource, TResult>(this Opt<TSource> source, Func<TSource, int, TResult> selector)
-        {
-            return source.HasValue ? Full(selector(source.ValueOrDefault, 0)) : Empty<TResult>();
-        }
-
-        // The proper response for Min() or Max() for an arbitrary unknown type.
-        // If the source is empty and the type is non-nullable, throw a No-Elements exception.
-        // If the source is empty and the type is nullable, return null.
-        // Otherwise, return the contained value (even if it is null).
-        private static TSource SingleOrNull<TSource>(this IOpt<TSource> source) => default(TSource) == null ? source.SingleOrDefault() : source.Single();
-
-        // Performs the later half of the long-form SelectManyRaw.
-        private static IOpt<TResult> SubSelect<TSource, TCollection, TResult>(TSource sourceElement, IOpt<TCollection> subCollection, Func<TSource, TCollection, TResult> resultSelector)
-        {
-            return subCollection.ToFixed().SelectRaw(subelement => resultSelector(sourceElement, subelement));
-        }
-
-        private static Opt<TSource> ToFixed<TSource>(IEnumerator<TSource> source)
+        private static Opt<TSource> Fix<TSource>(IEnumerator<TSource> source)
         {
             if (source.MoveNext())
             {
@@ -676,7 +212,7 @@ namespace Hgk.Zero.Options
             }
         }
 
-        private static Opt<object> ToFixed(IEnumerator source)
+        private static Opt<object> FixUntyped(System.Collections.IEnumerator source)
         {
             if (source.MoveNext())
             {
@@ -694,33 +230,6 @@ namespace Hgk.Zero.Options
             {
                 return Empty<object>();
             }
-        }
-
-        private static bool TryGetValueRaw<TSource>(this Opt<TSource> source, out TSource value)
-        {
-            value = source.ValueOrDefault;
-            return source.HasValue;
-        }
-
-        private static Opt<TSource> WhereNotNullRaw<TSource>(this Opt<TSource> source) =>
-            source.HasValue && source.ValueOrDefault != null ? source : Empty<TSource>();
-
-        private static Opt<TSource> WhereNotNullSelectValueRaw<TSource>(this Opt<TSource?> source) where TSource : struct =>
-            (source.HasValue && source.ValueOrDefault.HasValue) ? Full(source.ValueOrDefault.Value) : Empty<TSource>();
-
-        private static Opt<TSource> WhereNotRaw<TSource>(this Opt<TSource> source, Func<TSource, bool> predicate)
-        {
-            return (source.HasValue && !predicate(source.ValueOrDefault)) ? source : Empty<TSource>();
-        }
-
-        private static Opt<TSource> WhereNotRaw<TSource>(this Opt<TSource> source, Func<TSource, int, bool> predicate)
-        {
-            return (source.HasValue && !predicate(source.ValueOrDefault, 0)) ? source : Empty<TSource>();
-        }
-
-        private static Opt<TSource> WhereRaw<TSource>(this Opt<TSource> source, Func<TSource, int, bool> predicate)
-        {
-            return (source.HasValue && predicate(source.ValueOrDefault, 0)) ? source : Empty<TSource>();
         }
     }
 }
