+++ Hgk.Zero.Options/Try/TryFunction.cs
