+++ Hgk.Zero.Options/Try/TryToOpt.cs
